Update chat messages style
For aesthetic purposes, you need to compose concise messages in the chat and in the titles of dialogues also highlight places with key information in a different color. Therefore, this commit shortened most of the chat messages and some of the dialogue titles, and also added highlighting the player's markings.
In addition, to display Russian characters in the chat, you need to use the Windows-1257 encoding. This commit adds a UTF-8 to Windows-1257 conversion where it was omitted.
Otherwise, the program code has been cleaned up a little.